demo.c: time G_DoMoveThings() using a double instead of an int32_t.
I think this may fix some negative "non-profiled overhead" that I
have been seeing.

git-svn-id: https://svn.eduke32.com/eduke32@3182 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/demo.c
@@ -728,7 +728,7 @@ nextdemo_nomenu:
 
                 if (Demo_IsProfiling())
                 {
-                    int32_t t = (int32_t)gethitickms();
+                    double t = gethitickms();
                     G_DoMoveThings();
                     Demo_GToc(t);
                 }
